refactor(mac-settings): consolidate system model management cards
- replace the old experimental provider-testing and legacy model-section settings with a single `SettingsView+DictationModels.swift` surface
- add an `Active Model` card that keeps provider selection and per-model install state together in one card
- render separate Whisper and Parakeet model rows inside that card with shared install/remove/progress behavior
- move the `Active Model` card under the `SYSTEM` section and match its card spacing to the other multi-card settings groups
- split `Launch at Login` into its own standalone system card instead of combining it with model controls
- remove the old `SettingsView+ExperimentalProviderTesting.swift` and `SettingsView+ModelSection.swift` files
- update the macOS Xcode project to drop the removed settings files and include `SettingsView+DictationModels.swift`
